ai_service/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
pillow
requests
